2024-2-27 가게 수정

+++ src/main/java/com/green/gogiro/shop/ShopMapper.java
@@ -41,7 +41,7 @@ public interface ShopMapper {
 
     List<OwnerShopPicsProcVo> selByShopPics(int ishop);
 
-    int insFacilities(long ishop,List<Integer> ifacil);
+    int insFacilities(int ishop,List<Integer> ifacil);
 
     int delFacilities(long ishop);
